requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
opensearch-py==2.7.1
qdrant-client==1.9.2
sentence-transformers==3.2.1
transformers==4.44.2
torch>=2.1.0
peft==0.13.2
datasets==3.0.1
neo4j==5.23.1
ragas==0.1.15
numpy==1.26.4
requests==2.32.3
pydantic==2.8.2
openai==1.50.0
opencc-python-reimplemented==0.1.7
Pillow==10.4.0
pytesseract==0.3.10

# PaddleOCR: upgraded to 2.8+ to avoid PyMuPDF conflicts on Apple Silicon
paddleocr>=2.8.0
